split ui service from ui webapp, so we can provide alt implement
