+++ bistoury-ui-service-impl/src/main/java/qunar/tc/bistoury/ui/dao/impl/GitlabPrivateTokenDaoImpl.java
