+++ bistoury-ui-service-impl/src/main/java/qunar/tc/bistoury/ui/service/impl/GitlabPrivateTokenServiceImpl.java
