+++ bistoury-ui-service/src/main/java/qunar/tc/bistoury/ui/model/ReleaseInfo.java
