+++ bistoury-ui-service/src/main/java/qunar/tc/bistoury/ui/service/GitlabPrivateTokenService.java
