+++ bistoury-ui-service/src/main/java/qunar/tc/bistoury/ui/service/URLRedirectService.java
